+++ mrb_doc/models/image.rb
@@ -185,6 +185,15 @@ class Image
     self
   end
 
+  # Change the contrast of the image
+  # @param contrast [Float] a value between -100 and 100
+  # @raise [ArgumentError] If the contrast is outside of the allowed range
+  # @return [nil]
+  def contrast!(contrast)
+    # src/mruby_integration/models/image.cpp
+    self
+  end
+
   # Generates a new image of width by height in the specified colour.
   # @param width [Integer]
   # @param height [Integer]